requirements.txt
langchain
pyPDF2
python-dotenv
streamlit
langchain-community
openai
tiktoken
faiss-cpu